Improving KCLayout.delete

+++ src/kfactory/layout.py
@@ -198,7 +198,7 @@ class Factories[F: WrappedKCellFunc[Any, Any] | WrappedVKCellFunc[Any, Any]](
 
     def get_by_qualified_name(self, qualified_name: str) -> F | None:
         for factory in self._all:
-            if factory.qualified_name == qualified_name:
+            if factory.qualified_name == qualified_name:  # ty:ignore[invalid-attribute-access]
                 return factory
         return None
 
@@ -2816,8 +2816,11 @@ class KCLayout(
         return f"{self.__class__.__name__}({self.name}, n={len(self.kcells)})"
 
     def delete(self) -> None:
-        del kcls[self.name]
-        self.library.delete()
+        if kcls.get(self.name) is self:
+            del kcls[self.name]
+        if not self.library._destroyed():
+            self.library.delete()
+        self.tkcells = {}
 
     def routing_strategy(
         self,
